# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RelaySMS Publisher",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/creds": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "List credentials",
        "description": "Scope: creds:read.",
        "operationId": "list_credentials_v1_creds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialInfo"
                  },
                  "type": "array",
                  "title": "Response List Credentials V1 Creds Get"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": [
              "creds:read"
            ]
          },
          {
            "SessionCookie": [
              "creds:read"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Create credential",
        "description": "The generated password is shown only in this response. Scope: creds:write.",
        "operationId": "create_credential_v1_creds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialWithPassword"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username, an unknown scope, or a scope without the one it requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:write, or a cookie request came from an origin that isn't allowed. Also when granting scopes you don't hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "That username is already taken.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": [
              "creds:write"
            ]
          },
          {
            "SessionCookie": [
              "creds:write"
            ]
          }
        ]
      }
    },
    "/v1/creds/{username}": {
      "get": {
        "tags": [
          "Credentials"
        ],
        "summary": "Get credential",
        "description": "Returns the ETag to send as If-Match. Scope: creds:read.",
        "operationId": "get_credential_v1_creds__username__get",
        "security": [
          {
            "HTTPBasic": [
              "creds:read"
            ]
          },
          {
            "SessionCookie": [
              "creds:read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialInfo"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No credential with that username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Credentials"
        ],
        "summary": "Update credential",
        "description": "Deactivating ends its sessions. Needs If-Match. Scope: creds:write.",
        "operationId": "update_credential_v1_creds__username__patch",
        "security": [
          {
            "HTTPBasic": [
              "creds:write"
            ]
          },
          {
            "SessionCookie": [
              "creds:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Username"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialInfo"
                }
              }
            }
          },
          "400": {
            "description": "Nothing to change, an unknown scope, or a scope without the one it requires.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:write, or a cookie request came from an origin that isn't allowed. Also when changing your own credential or one with scopes you don't hold, or granting scopes you don't hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No credential with that username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Changed by another request at the same moment. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "If-Match isn't the current ETag: it changed since you loaded it. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "No If-Match header. Send the ETag from the last GET.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Credentials"
        ],
        "summary": "Delete credential",
        "description": "Needs If-Match. Scope: creds:write.",
        "operationId": "delete_credential_v1_creds__username__delete",
        "security": [
          {
            "HTTPBasic": [
              "creds:write"
            ]
          },
          {
            "SessionCookie": [
              "creds:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Username"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:write, or a cookie request came from an origin that isn't allowed. Also when changing your own credential or one with scopes you don't hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No credential with that username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Changed by another request at the same moment. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "If-Match isn't the current ETag: it changed since you loaded it. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "No If-Match header. Send the ETag from the last GET.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/creds/{username}/reset-password": {
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Reset password",
        "description": "New password shown once; sessions end. Needs If-Match. Scope: creds:write.",
        "operationId": "reset_credential_password_v1_creds__username__reset_password_post",
        "security": [
          {
            "HTTPBasic": [
              "creds:write"
            ]
          },
          {
            "SessionCookie": [
              "creds:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialWithPassword"
                }
              }
            }
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:write, or a cookie request came from an origin that isn't allowed. Also when changing your own credential or one with scopes you don't hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No credential with that username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Changed by another request at the same moment. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "412": {
            "description": "If-Match isn't the current ETag: it changed since you loaded it. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "428": {
            "description": "No If-Match header. Send the ETag from the last GET.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/creds/{username}/revoke-sessions": {
      "post": {
        "tags": [
          "Credentials"
        ],
        "summary": "Revoke sessions",
        "description": "Logs it out everywhere; Basic auth still works. Scope: creds:write.",
        "operationId": "revoke_credential_sessions_v1_creds__username__revoke_sessions_post",
        "security": [
          {
            "HTTPBasic": [
              "creds:write"
            ]
          },
          {
            "SessionCookie": [
              "creds:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32,
              "title": "Username"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No session cookie or Basic credentials, an expired session, or a wrong username or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The credential lacks creds:write, or a cookie request came from an origin that isn't allowed. Also when changing your own credential or one with scopes you don't hold.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No credential with that username.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Changed by another request at the same moment. Reload and try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A parameter or body field is missing or invalid; the message names it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CredentialCreate": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 32,
            "title": "Username"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "minItems": 1,
            "title": "Scopes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "scopes"
        ],
        "title": "CredentialCreate"
      },
      "CredentialInfo": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "administrator": {
            "type": "boolean",
            "title": "Administrator",
            "description": "Holds every scope."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "active_sessions": {
            "type": "integer",
            "title": "Active Sessions"
          }
        },
        "type": "object",
        "required": [
          "username",
          "active",
          "scopes",
          "administrator",
          "created_at",
          "active_sessions"
        ],
        "title": "CredentialInfo"
      },
      "CredentialUpdate": {
        "properties": {
          "scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 32,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Scopes"
          },
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CredentialUpdate"
      },
      "CredentialWithPassword": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "administrator": {
            "type": "boolean",
            "title": "Administrator",
            "description": "Holds every scope."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "active_sessions": {
            "type": "integer",
            "title": "Active Sessions"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Shown only in this response."
          }
        },
        "type": "object",
        "required": [
          "username",
          "active",
          "scopes",
          "administrator",
          "created_at",
          "active_sessions",
          "password"
        ],
        "title": "CredentialWithPassword"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "What went wrong, safe to show to users."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The body of every error response."
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "SessionCookie": {
        "type": "apiKey",
        "description": "Set by POST /v1/auth/login.",
        "in": "cookie",
        "name": "relaysms_session"
      }
    }
  }
}